user authentication, login logout, user can upload image as a post, basic functionality completed

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,19 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -43,13 +56,36 @@ class PostsController extends Controller
     {
         $this->validate($request, [
                 'title' => 'required',
-                'body' => 'required'
+                'body' => 'required',
+                'cover_image' => 'image|nullable|max:1999'
             ]);
+
+        //Handle File Upload
+        if($request->hasFile('cover_image')){
+            //get filename with extention
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }else{
+            $fileNameToStore = 'noimage.jpg';   //default image
+        }
 
         //create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save(); 
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -76,6 +112,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        //Check for correct user
+        if(auth()->user()->id != $post->user_id){
+             return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -93,6 +133,24 @@ class PostsController extends Controller
                 'title' => 'required',
                 'body' => 'required'
             ]);
+
+         //Handle File Upload
+        if($request->hasFile('cover_image')){
+            //get filename with extention
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            //get just file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            //get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            //filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
 
         //Edit Post
         $post = Post::find($id);
@@ -112,6 +170,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+         if(auth()->user()->id != $post->user_id){
+             return redirect('/posts')->with('error', 'Unauthorized Page');
+         }
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
